Dataset and 'not null' constraints added in 'create_data_model_view'.
