+++ tests/algorithm_validation_tests/test_pearson_validation.py
@@ -16,30 +16,7 @@ def pearson_request(input):
     if input["inputdata"]["x"]:
         variables.extend(input["inputdata"]["x"])
 
-    filters = {
-        "condition": "AND",
-        "rules": [
-            {
-                "id": "dataset",
-                "type": "string",
-                "value": input["inputdata"]["datasets"],
-                "operator": "in",
-            },
-            {
-                "condition": "AND",
-                "rules": [
-                    {
-                        "id": variable,
-                        "type": "string",
-                        "operator": "is_not_null",
-                        "value": None,
-                    }
-                    for variable in variables
-                ],
-            },
-        ],
-        "valid": True,
-    }
+    filters = None
     input["inputdata"]["filters"] = filters
     request_json = json.dumps(input)
