+++ tests/standalone_tests/__init__.py
@@ -1,0 +1,1 @@
+algorithms_url = "http://127.0.0.1:5000/algorithms"
