+++ tests/standalone_tests/nodes_communication_helper.py
@@ -1,42 +1,3 @@
-from os import path
-
-import toml
-from celery import Celery
-
-from mipengine import AttrDict
-from tests.standalone_tests.conftest import TEST_ENV_CONFIG_FOLDER
-
-
-def get_node_config_by_id(node_config_file: str):
-    with open(path.join(TEST_ENV_CONFIG_FOLDER, node_config_file)) as fp:
-        node_config = AttrDict(toml.load(fp))
-    return node_config
-
-
-def get_celery_app(node_config: AttrDict):
-    rabbitmq_credentials = (
-        node_config.rabbitmq.user + ":" + node_config.rabbitmq.password
-    )
-    rabbitmq_socket_addr = (
-        node_config.rabbitmq.ip + ":" + str(node_config.rabbitmq.port)
-    )
-    vhost = node_config.rabbitmq.vhost
-
-    return Celery(
-        "mipengine.node",
-        broker=f"amqp://{rabbitmq_credentials}@{rabbitmq_socket_addr}/{vhost}",
-        backend="rpc://",
-        include=[
-            "mipengine.node.tasks.tables",
-            "mipengine.node.tasks.remote_tables",
-            "mipengine.node.tasks.merge_tables",
-            "mipengine.node.tasks.views",
-            "mipengine.node.tasks.udfs",
-            "mipengine.node.tasks.common",
-        ],
-    )
-
-
 def get_celery_task_signature(celery_app, task):
     signature_mapping = {
         "get_node_info": celery_app.signature(
